Reading departure delays only if arrival delays are not provided
GTFS realtime updaters can have different values for arrival delay and
departure delay for a same stop on the same trip. A departure delay
smaller than an arrival delay (or greater, if delays are negative) can
cause a departure scheduled before the arrival. In this case the
real time validator discards teh update. For most city bus updaters this
is not the desired behavior. In particular the CTT Nord rt updaters
have this issue.
Anyway, this is just a temporary patch. The choice of read or not the
departure dalays must be made configurable with a property in
router-config.json.